LayoutTests/http/tests/inspector/network/network-xhr-same-url-as-main-resource-expected.txt

CONSOLE MESSAGE: line 13: Done.
Tests that XHRs with the same url as a main resource have correct category.

Bug 68648
xhr
